refactor: clarify the tally report types
This changes the types for the candidate and yesno tallies from arrays into objects. For candidate contests, it splits the former single array that contained both write-ins and ballot candidates into separate properties on an object. For yesno contests, instead of a 2-element tuple we now have an object with "yes" and "no" properties.

I think this is simpler and requires less casting. It also adds a few assertions for situations which shouldn't happen but otherwise might silently fail.