refactor: prepare printer for PDF printing
This doesn't change any behavior for now, but it adds the concept of a print job type. For now, the only print job type we're using is `CurrentPage` which will print the current page. In the future, we'll use the `PDFMakeDocument` type which `module-print` will accept but the fallback "local" printer won't. This allows us to determine whether we can print PDFs and, if not, fall back to printing using `CurrentPage` type.