refactor: inject card and storage APIs into app
This allows both `App` and `AppRoot` to be agnostic to whether they're running in the sample app or not. The basic idea is that the app works through mocked card and storage objects in the sample app and real ones otherwise.

This will make it easier to, in the future, move away from local storage entirely and onto a backend service. Our tests still assume local storage, but they can be migrated to be agnostic as well.