style(lint): disallow `null` with a custom error message
This brings back the functionality of the `no-null` eslint pugin using the built-in `no-restricted-syntax` rule. See the selector docs here: https://eslint.org/docs/developer-guide/selectors and the rule docs here: https://eslint.org/docs/rules/no-restricted-syntax.

The downside is that we can't ignore _only_ the "no null" part of `no-restricted-syntax`, but I don't think it'll be a problem. If is is, we can revisit this with a custom rule that'd be pretty trivial.