fix: don't allow double-tapping write-in "Accept"
This was only reproducible with touch events (i.e. tapping) rather than click events. The fix is simply to remove the timeout that we previously needed but no longer do after the changes made to fix tap/click handling a while ago. We could also have added an explicit guard, but as this is now synchronous I didn't deem it worth the effort.

Fixes #1028