refactor: simplify printer API
This removes a lot of the speculative work I previously did to support printing via module-print, print PDFs, etc. Since we're now planning to just use Electron's printing of the current page, much of this is not needed or is not a good fit. Print status may be added back in the future, but in a different form.

In addition, this replaces the `fetchMock`-based testing of printing and replaces it with dependency-injected fake printers. This is the same approach used for the memory card access and persistent storage.