fix: ensure speak-after-cancel works
On Linux, when a call to `speechSynthesis.speak(utterance)` is immediately preceded by `speechSynthesis.cancel()`, then `utterance` will not be spoken out loud. To work around that, we issue an empty utterance to be sacrificed at the altar of Linux + speech-dispatcher + Chromium so that we all may live… I mean, so `utterance` will be spoken properly.