test: remove remaining jest timeout increases
This is more or less the same as what I did with 2fa8a6c. I didn't spend much time trying to understand why I couldn't replace _all_ calls to `advanceTimers` with `await advanceTimersAndPromises`, but suffice it to say that the tests failed if I did in the places where I left them alone. There are still gremlins lurking in here that I'd like to clean out when we have more time.