fix: do not print multiple ballots
I believe the polling was the cause of the multiple ballots being printed, so this commit removes the polling. Instead, it only triggers a printer refresh when _any USB device (that isn't the accessible controller) is added or removed_. The reason polling caused so many ballots to be printed was that `kiosk-browser`'s `getPrinterInfo` API was calling out to `lpinfo -v`, which for some reason _does not include the Brother printer when it's actually printing_. This may or may not be true with other printers, but it does _not_ seem to be true with the Parallels "Print to Mac (PDF)" virtual printer. Yay testing with real hardware.

The changes to `MemoryHardware` to support this were to make printer status be like adding or removing a device, similar to the accessible controller. Also, this commit includes information about the Brother printer we use, but it's not tied to that printer and should work with anything that we've configured properly at the `lpadmin` layer.

Closes #1112