feat: monitor card reader via USB device changes
This replaces the polling of `/card/reader` with a device add notification that checks for the specific card reader we use in production. The upside is that plugging or unpluging triggers updates immediately. The main downside is that this hardcodes a specific product rather than being generalized to any card reader that psycard supports. I think the trade-off is worth it, but I'm open to other opinions.