fix: ensure "Card Reader Not Detected" shows as expected
Previously, we were checking `/card/reader` on USB events that didn't match a known device (which was just the accessible controller). With this change, we now look for the card reader too but by device & manufacturer name rather than vendor ID and product ID. @benadida was concerned that those values might not be stable for all the devices we have, but that the string names would be.

Avoiding `/card/reader` has the benefit of avoiding a race condition that rarely happened in development but seems to always happen on the real hardware. BMD would receive a USB event and request `/card/reader` before `module-smartcards` ever noticed the card reader was connected, so the request would return `{ connected: false }`. This caused a persistent "Card Reader Not Detected" screen. With this fix it should appear only when the card reader is actually absent.