refactor: introduce `globalThis`
`globalThis` was introduced a while back to solve the "how do I reference the global object?" problem. It's implemented in Chrome, Safari, Node.js, Firefox, etc. I'm adding it only to the `setupTests.tsx` file for now because `globalThis` is _not_ currently supported in the default browser for Windows 10, the legacy version of Microsoft Edge (i.e. not the new version based on Chrome). If we care about supporting legacy Microsoft Edge, we can easily add a polyfill.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/globalThis