test: update tests to match "new" election definition
In https://github.com/votingworks/ballot-encoder/commit/cfa65006ff21214803d07fd4c656f21b9842c1b1 the sample election was changed to shrink the number of candidates in the Governor contest. Updating ballot-encoder finally pulled this change into bmd, so the test now needs to change to reflect that the contest only has 16 candidates and not 26.